refactor: change f32 to f64 in hittable.zig, ray.zig, image.zig, interval.zig, and vec3.zig for better precision refactor: change Vec3 struct in vec3.zig to use an array for coordinates for better code readability and maintainability refactor: update Vec3 methods in vec3.zig to use new array-based coordinates refactor: update tests in vec3.zig to reflect changes in Vec3 struct and methods

+++ src/entities/hittable.zig
@@ -14,7 +14,7 @@ const HittableList = hittableList.HittableList;
 pub const HitRecord = struct {
     p: Vec3,
     normal: Vec3,
-    t: f32,
+    t: f64,
     front_face: bool = false,
 
     const Self = @This();
